Renamed the token_context module.
