+++ crates/bld_runner/src/artifacts/version2.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     pipeline::traits::{ApplyTokens, CompleteTokenTransformer},
-    token_context::version2::PipelineContext,
+    token_context::v2::PipelineContext,
 };
 
 #[derive(Debug, Default, Serialize, Deserialize)]
