+++ crates/bld_runner/src/external/version2.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 
 use crate::{
     pipeline::traits::{ApplyTokens, CompleteTokenTransformer},
-    token_context::version2::PipelineContext,
+    token_context::v2::PipelineContext,
 };
 
 #[derive(Debug, Default, Serialize, Deserialize)]
