+++ crates/bld_runner/src/pipeline/version2.rs
@@ -2,7 +2,7 @@ use crate::artifacts::version2::Artifacts;
 use crate::external::version2::External;
 use crate::platform::version2::Platform;
 use crate::step::version2::BuildStep;
-use crate::token_context::version2::PipelineContext;
+use crate::token_context::v2::PipelineContext;
 use anyhow::Result;
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
