+++ crates/bld_runner/src/step/version2.rs
@@ -8,7 +8,7 @@ use std::path::PathBuf;
 
 use crate::{
     pipeline::traits::{ApplyTokens, CompleteTokenTransformer},
-    token_context::version2::PipelineContext,
+    token_context::v2::PipelineContext,
 };
 
 #[derive(Debug, Default, Serialize, Deserialize)]
